Added dialog messages for feedback on GUI

+++ BarberProgram/src/controllersPackage/ServicesTabController.java
@@ -11,7 +11,9 @@ import java.net.URL;
 import java.net.URLConnection;
 import java.net.URLEncoder;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
+import java.util.concurrent.Callable;
 
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -20,19 +22,28 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.stage.Stage;
 import mainPackage.Connection;
 import mainPackage.Service;
 import mainPackage.User;
 
 /**
  * Controller for the services tabbed page
+ * 
  * @author Raj
  */
 public class ServicesTabController implements Initializable {
@@ -72,25 +83,87 @@ public class ServicesTabController implements Initializable {
 		ObservableList<String> list = FXCollections.observableArrayList(cities);
 		serviceChoiceBox.setItems(list);
 	}
+	
+	private void createDialog(String msg, int btnAmount, String[] btnNames, Callable[] func){
+		// Logout Alert Window
+		Alert alert = new Alert(AlertType.NONE);
+		alert.setContentText(msg);
+		
+		ButtonType[] btns = new ButtonType[btnAmount];
+		Callable[] functions = new Callable[btnAmount];
+		
+		// used to check if we passed any functions to this dialog
+		boolean nofunc = false;
+		
+		if(func != null){
+			for(int i=0; i<btnAmount; i++){
+				functions[i] = func[i];	
+			}
+		}else{
+			// Set the flag to false;
+			nofunc = true;
+		}
+		
+		for(int i=0; i<btnAmount; i++){
+			btns[i] = new ButtonType(btnNames[i]);
+		}
+
+		alert.getButtonTypes().setAll(btns);
+
+		Optional<ButtonType> result = alert.showAndWait();
+
+		
+		for(int i=0; i<btnAmount; i++){
+			if (result.get() == btns[i]) {
+				if(!nofunc){
+					try {
+						func[i].call();
+					} catch (Exception e) {
+						e.printStackTrace();
+					}
+				}
+			}else{
+				
+			}
+		}
+		
+	}
 
 	@FXML
 	protected void handleRemoveService(ActionEvent event) throws IOException {
 
-		// Make sure something is selected
-		try {
-			Service selectedItems = serviceTable.getSelectionModel().getSelectedItems().get(0);
-			String removeService = selectedItems.toString();
-			System.out.println(removeService);
+		Service selectedItems = serviceTable.getSelectionModel().getSelectedItems().get(0);
+		
+		if(selectedItems == null){
+			createDialog("Select something first", 1, new String[]{"Ok"}, null);
+		}else{
+			// Logout Alert Window
+			Alert alert = new Alert(AlertType.NONE);
+			alert.setContentText("Are you sure you want to delete this service?");
+			ButtonType buttonTypeOne = new ButtonType("Yes");
+			ButtonType buttonTypeTwo = new ButtonType("No");
+			alert.getButtonTypes().setAll(buttonTypeOne, buttonTypeTwo);
+			Optional<ButtonType> result = alert.showAndWait();
 
-			// Send this value to a php script
-			if (deleteService(selectedItems.getId())) {
-				// Remove from Users services Array
-				removeFromUsers(selectedItems);
-				// After we get confirmation from the script, remove from GUI
-				updateGUI();
+			if (result.get() == buttonTypeOne) {
+				// ReMake sure something is selected
+				try {
+					selectedItems = serviceTable.getSelectionModel().getSelectedItems().get(0);
+					String removeService = selectedItems.toString();
+					System.out.println(removeService);
+
+					// Send this value to a php script
+					if (deleteService(selectedItems.getId())) {
+						// Remove from Users services Array
+						removeFromUsers(selectedItems);
+						// After we get confirmation from the script, remove from
+						// GUI
+						updateGUI();
+					}
+				} catch (NullPointerException e) {
+					System.err.println("Select something first.");
+				}
 			}
-		} catch (NullPointerException e) {
-			System.err.println("Select something first.");
 		}
 	}
 
@@ -183,7 +256,7 @@ public class ServicesTabController implements Initializable {
 
 		return deleted;
 	}
-	
+
 	@FXML
 	protected void handleSubmitService(ActionEvent event) throws IOException {
 
@@ -192,6 +265,7 @@ public class ServicesTabController implements Initializable {
 		// Validate that data was selected/entered (service name, price)
 		if (serviceChoiceBox.getValue() == null || tfPrice.getText().equals("")) {
 			valid = false;
+			createDialog("Either service or price is not entered!", 1, new String[]{"Ok"}, null);
 		} else {
 			// Both have been filled
 
@@ -199,8 +273,8 @@ public class ServicesTabController implements Initializable {
 			for (int i = 0; i < User.getInstance().services.size(); i++) {
 				if (serviceChoiceBox.getValue().equals(User.getInstance().services.get(i).getService())) {
 					System.out.println("This service already exists! ");
-					valid = false;
-					break;
+					createDialog("This service already exists!", 1, new String[]{"Ok"}, null);
+					return;
 				}
 			}
 
@@ -208,7 +282,8 @@ public class ServicesTabController implements Initializable {
 			try {
 				double price = Double.parseDouble(tfPrice.getText());
 			} catch (NumberFormatException e) {
-				valid = false;
+				createDialog("Incorrect price entered!", 1, new String[]{"Ok"}, null);
+				return;
 			}
 		}
